count(config) return int

+++ src/client/headers/client/sqf/config.hpp
@@ -42,7 +42,7 @@ namespace intercept {
         std::vector<config> config_properties(const config &config_entry, sqf_string_const_ref condition_ = "true", bool inherit = true);
         sqf_return_string config_source_mod(const config &config_entry_);
         sqf_return_string_list config_source_mod_list(const config &config_entry_);
-        float count(const config &config_entry_);
+        int count(const config &config_entry_);
         //std::vector<game_value> get_array(const config &config_entry_);
         config get_mission_config(sqf_string_const_ref value_);
         float get_number(const config &config_entry_);
